+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added Configuration option to hide "Drop Item" message in chat (hidden by default).
+
 ## [1.0.5] - 2022-07-30
 
 - Added Configuration option to forcefully enable easter eggs (under Developer Settings).